sd_tarea_1/qa_score/requirements.txt
google-generativeai==0.8.5
fastapi==0.115.0
uvicorn[standard]==0.30.6